+++ plugin-gradle/src/test/java/com/diffplug/gradle/spotless/BiomeIntegrationTest.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2023 DiffPlug
+ * Copyright 2023-2024 DiffPlug
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -339,5 +339,33 @@ class BiomeIntegrationTest extends GradleIntegrationHarness {
 		assertFile("biome_test.js").sameAsResource("biome/js/fileBefore.js");
 		assertThat(spotlessApply.getOutput()).contains("Format with errors is disabled.");
 		assertThat(spotlessApply.getOutput()).contains("Step 'biome' found problem in 'biome_test.js'");
+	}
+
+	/**
+	 * Biome is hard-coded to ignore certain files, such as package.json. Since version 1.5.0,
+	 * the biome CLI does not output any formatted code anymore, whereas previously it printed
+	 * the input as-is. This tests checks that when the biome formatter outputs an empty string,
+	 * the contents of the file to format are used instead.
+	 *
+	 * @throws Exception When a test failure occurs.
+	 */
+	@Test
+	void preservesIgnoredFiles() throws Exception {
+		setFile("build.gradle").toLines(
+				"plugins {",
+				"    id 'com.diffplug.spotless'",
+				"}",
+				"repositories { mavenCentral() }",
+				"spotless {",
+				"    json {",
+				"        target '**/*.json'",
+				"        biome('1.5.0')",
+				"    }",
+				"}");
+		setFile("package.json").toResource("biome/json/packageBefore.json");
+
+		var spotlessApply = gradleRunner().withArguments("--stacktrace", "spotlessApply").build();
+		assertThat(spotlessApply.getOutput()).contains("BUILD SUCCESSFUL");
+		assertFile("package.json").sameAsResource("biome/json/packageAfter.json");
 	}
 }